Improve React app entrypoint
